Move's Clerk to correct place, fixes middleware
