+++ apps/web/app/(authenticated)/auth/sign-up/email-signup.tsx
