+++ apps/web/app/(authenticated)/auth/sso-callback/[[...sso-callback]]/page.tsx
